// @types/FiiPapelCalculatedDataType.d.ts
export type FiiPapelCalculatedDataType = {
  ticker: string;
  category: string;
  price: number;
  dy: number;
  pvp: number;
  caixa: number;
  cagrDividendos3Anos: number;
  cagrValorCota3Anos: number;
  patrimonio: number;
  gestor: string;
  isTopManager: boolean;
  gestao: string;
  cotistas: number;
  liquidezDiaria: number;
  subcategoria: string;
};