// @types/FiiPapelFormattedDataType.d.ts
export type FiiPapelFormattedDataType = {
  ticker: string;
  price: string;
  category: string;
  dy: string;
  pvp: string;
  caixa: string;
  cagrDividendos3Anos: string;
  cagrDividendos3AnosColor: string;
  cagrValorCota3Anos: string;
  cagrValorCota3AnosColor: string;
  patrimonio: string;
  gestor: string;
  isTopManager: string;
  gestao: string;
  cotistas: string;
  liquidezDiaria: string;
  subcategoria: string;
};